+++ common/src/main/kotlin/com/cobblemon/mod/common/block/entity/PokemonPastureBlockEntity.kt
@@ -18,12 +18,14 @@ import com.cobblemon.mod.common.api.scheduling.afterOnServer
 import com.cobblemon.mod.common.api.text.red
 import com.cobblemon.mod.common.block.PastureBlock
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity
+import com.cobblemon.mod.common.net.messages.client.effect.SpawnSnowstormEntityParticlePacket
 import com.cobblemon.mod.common.net.messages.client.pasture.ClosePasturePacket
 import com.cobblemon.mod.common.net.messages.client.pasture.OpenPasturePacket
 import com.cobblemon.mod.common.net.messages.client.pasture.PokemonPasturedPacket
 import com.cobblemon.mod.common.net.serverhandling.storage.SendOutPokemonHandler
 import com.cobblemon.mod.common.pokemon.Pokemon
 import com.cobblemon.mod.common.util.DataKeys
+import com.cobblemon.mod.common.util.cobblemonResource
 import com.cobblemon.mod.common.util.lang
 import com.cobblemon.mod.common.util.server
 import com.cobblemon.mod.common.util.toVec3d
@@ -152,8 +154,14 @@ class PokemonPastureBlockEntity(pos: BlockPos, state: BlockState) :
                 entity.setPos(fixedPosition.center.subtract(0.0, 0.5, 0.0))
                 val pc = Cobblemon.storage.getPC(player)
                 entity.beamMode = 2
-                afterOnServer(seconds = SendOutPokemonHandler.SEND_OUT_DURATION) {
-                    entity.beamMode = 0
+                afterOnServer(seconds = 0.5F) {
+                    if(entity.pokemon.shiny) {
+                        SpawnSnowstormEntityParticlePacket(cobblemonResource("shiny_ring"), entity.id, listOf("shiny_particles", "middle"))
+                            .sendToPlayersAround(entity.x, entity.y, entity.z, 64.0, entity.level().dimension())
+                    }
+                    entity.after(seconds = SendOutPokemonHandler.SEND_OUT_DURATION) {
+                        entity.beamMode = 0
+                    }
                 }
                 if (world.addFreshEntity(entity)) {
                     val tethering = Tethering(